Move Wizard into next-specific directory for now
